Improving refactoring of API following @jonnytest1 feedback
