--- front/src/registered_callbacks.ts
@@ -1,4 +1,0 @@
-import type {IframeResponseEventMap} from "./Api/Events/IframeEvent";
-import type {IframeCallback} from "./Api/iframe/IframeApiContribution";
-
-export const registeredCallbacks: { [K in keyof IframeResponseEventMap]?: IframeCallback<K> } = {}
